refatora endpoint de sign in e sign up para colocar os handlers em outro arquivos
