+++ back-end-dart/salvaguardas_backend/bin/api/endpoint.dart
@@ -1,0 +1,5 @@
+import 'package:shelf/shelf.dart';
+
+abstract class EndPoint {
+  Handler get handler;
+}
