+++ back-end-dart/salvaguardas_backend/bin/main.dart
@@ -1,5 +1,5 @@
 import 'package:shelf/shelf.dart';
-import 'api/auth_api.dart';
+import 'api/auth/auth_api.dart';
 import 'api/user_api.dart';
 import 'api/workload_api.dart';
 import 'infra/custom_server.dart';
